move question part class to separate file
